% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asg_commonln5.R
\name{asg_commonln5}
\alias{asg_commonln5}
\title{asg_commonln5}
\usage{
asg_commonln5(y, x, count, group, priors, niter = 2000, nchains = 3,
  nclusters = nchains, burnin = niter/2, thin = 10, inits = NULL,
  mle = FALSE)
}
\arguments{
\item{y}{response variable which follows binomial dist}

\item{x}{explanatory variable}

\item{count}{n in binomial dist}

\item{group}{groups of response}

\item{priors}{list of priors}

\item{niter}{number of interations to be run (default=2000)}

\item{nchains}{number of chains to be run (default=3)}

\item{nclusters}{number of clusters to be used (default=nchains)}

\item{burnin}{number of samples to be used as burnin (technically adaption, see link below)}

\item{thin}{when you want to thin (default=10)}

\item{inits}{Add specific initial values}

\item{mle}{expects TRUE/FALSE, if TRUE will use maximum likelihood to get starting values. Needs 3 chains.}
}
\value{
A MCMC object
}
\description{
Runs Asymmetric Gaussian MCMC with a common mean structure accross the groups with log-normal priors with one intercept
}
\examples{
  priors$vb1
  priors$mn
  priors$vn
  priors$mx
  priors$vm
  priors$ms1
  priors$vs1
  priors$ms2
  priors$vs2

}
\seealso{
\url{http://www.mikemeredith.net/blog/2016/Adapt_or_burn.htm}
}
