% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asg_indep.R
\name{asg_indep}
\alias{asg_indep}
\title{asg_indep}
\usage{
asg_indep(y, x, count, group, priors, niter, nchains = 3, burnin = niter/10,
  thin = 10)
}
\arguments{
\item{y}{response variable which follows binomial dist}

\item{x}{explanatory variable}

\item{count}{n in binomial dist}

\item{group}{groups of response}

\item{priors}{list of priors}

\item{niter}{number of interations to be run}

\item{nchains}{number of chains to be run (default=3)}

\item{burnin}{number of samples to be used as burnin (technically adaption, see link below)}

\item{thin}{when you want to thin (default=10)}
}
\value{
A MCMC object
}
\description{
Runs Asymmetric Gaussian MCMC with an independent mean structure accross the groups
}
\examples{
priors <- list()
priors$vb1
priors$vb2
priors$ve
priors$mx
priors$vm
priors$vs1
priors$vs2


}
\seealso{
\url{http://www.mikemeredith.net/blog/2016/Adapt_or_burn.htm}
}
