% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asg_indep5.R
\name{asg_indep5}
\alias{asg_indep5}
\title{asg_indep5}
\usage{
asg_indep5(y, x, count, group, priors, niter = 2000, nchains = 3,
  nclusters = nchains, burnin = niter/2, thin = 10)
}
\arguments{
\item{y}{response variable which follows binomial dist}

\item{x}{explanatory variable}

\item{count}{n in binomial dist}

\item{group}{groups of response}

\item{priors}{list of priors}

\item{niter}{number of interations to be run (default=2000)}

\item{nchains}{number of chains to be run (default=3)}

\item{nclusters}{number of clusters to be used (default=nchains)}

\item{burnin}{number of samples to be used as burnin (technically adaption, see link below)}

\item{thin}{when you want to thin (default=10)}
}
\value{
A MCMC object
}
\description{
Runs Asymmetric Gaussian MCMC with an independent mean structure accross the groups with only one intercept
}
\examples{
priors <- list()
priors$vb1
priors$mn
priors$vn
priors$mx
priors$vm
priors$vs1
priors$vs2


}
\seealso{
\url{http://www.mikemeredith.net/blog/2016/Adapt_or_burn.htm}
}
