% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_finder.R
\name{asg_mle}
\alias{asg_mle}
\title{mle_finder}
\usage{
asg_mle(y, x, num, group, season, inits = NULL, ll = NULL)
}
\arguments{
\item{y}{response variable which follows binomial dist}

\item{x}{explanatory variable}

\item{group}{groups of response}

\item{season}{seasons in response}

\item{inits}{initial values for optim (defualt=NULL)}

\item{ll}{likelihood to be optimized (defualt=ASG)}

\item{count}{n in binomial dist}
}
\value{
A function
}
\description{
A function which finds and formats the mle of ASG ff
}
\examples{
asg_mle(y, x, num, group, seas)

}
