% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asg_multi.R
\name{asg_multi}
\alias{asg_multi}
\title{asg_multi}
\usage{
asg_multi(y, x, count, group, priors, niter, nchains = 3, burnin = niter/10,
  thin = 10)
}
\arguments{
\item{y}{response variable which follows binomial dist}

\item{x}{explanatory variable}

\item{count}{n in binomial dist}

\item{group}{groups of response}

\item{priors}{list of priors}

\item{niter}{number of interations to be run}

\item{nchains}{number of chains to be run (default=3)}

\item{burnin}{number of samples to be used as burnin (technically adaption, see link below)}

\item{thin}{when you want to thin (default=10)}
}
\value{
A MCMC object
}
\description{
Runs Asymmetric Gaussian MCMC with an multivariate, independent mean structure accross the groups
}
\examples{
mean.vector = c(beta1, beta2, leta, mu, lsigma1, lsigma2)

priors <- list()
priors$mu0  # A 6x1 vector
priors$C    # A 6x6 precision matrix for mean MVN
priors$phi0 # A 6x6 precision matrix for Wish Dist


}
\seealso{
\url{http://www.mikemeredith.net/blog/2016/Adapt_or_burn.htm}
}
