% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_res_df.R
\name{get_res_df}
\alias{get_res_df}
\title{get_res_df}
\usage{
get_res_df(dat, x, niter, model.name)
}
\arguments{
\item{dat}{data in same format as used in functions}

\item{x}{mcmc object}

\item{niter}{number of interations to be ran}

\item{model.name}{name of the model associated with df}
}
\value{
A data.frame
}
\description{
Use mcmc objects from package to get posterior estimates of theta
}
\examples{




}
