% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asg_hierln.R
\name{asg_hierln}
\alias{asg_hierln}
\title{asg_hierln}
\usage{
asg_hierln(y, x, count, group, priors, niter = 2000, nchains = 3,
  nclusters = nchains, burnin = niter/2, thin = 10, inits = NULL,
  mle = FALSE)
}
\arguments{
\item{y}{response variable which follows binomial dist}

\item{x}{explanatory variable}

\item{count}{n in binomial dist}

\item{group}{groups of response}

\item{priors}{list of priors}

\item{niter}{number of interations to be run (default=2000)}

\item{nchains}{number of chains to be run (default=3)}

\item{nclusters}{number of clusters to be used (default=nchains)}

\item{burnin}{number of samples to be used as burnin (technically adaption, see link below) (default=niter/2)}

\item{thin}{when you want to thin (default=10)}

\item{inits}{Add specific initial values}

\item{mle}{expects TRUE/FALSE, if TRUE will use maximum likelihood to get starting values. Needs 3 chains.}
}
\value{
A MCMC object
}
\description{
Runs Asymmetric Gaussian MCMC with a hierarchical mean structure accross the groups using log-normal hierarchical models
}
\examples{
priors = list()
#betas
priors$vtb1
priors$vtb2
# etas
priors$mn
priors$vmn
priors$vtn
# mus
priors$mx
priors$vmx
priors$vtm
# sigmas
priors$ms1
priors$vms1
priors$ms2
priors$vms2
priors$vts1
priors$vts2


}
\seealso{
\url{http://www.mikemeredith.net/blog/2016/Adapt_or_burn.htm}
}
