% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asg_hierln.R
\name{asg_hierln}
\alias{asg_hierln}
\title{asg_hierln}
\usage{
asg_hierln(y, x, count, group, priors, niter, nchains = 3,
  burnin = niter/10, thin = 10)
}
\arguments{
\item{y}{response variable which follows binomial dist}

\item{x}{explanatory variable}

\item{count}{n in binomial dist}

\item{group}{groups of response}

\item{priors}{list of priors}

\item{niter}{number of interations to be run}

\item{nchains}{number of chains to be run (default=3)}

\item{burnin}{number of samples to be used as burnin (technically adaption, see link below)}

\item{thin}{when you want to thin (default=10)}
}
\value{
A MCMC object
}
\description{
Runs Asymmetric Gaussian MCMC with a hierarchical mean structure accross the groups using log-normal hierarchical models
}
\examples{
priors = list()
priors$vtb1
priors$vtb2
priors$vte
priors$mx
priors$vmx
priors$vtm
priors$vts1
priors$vts2


}
\seealso{
\url{http://www.mikemeredith.net/blog/2016/Adapt_or_burn.htm}
}
