% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drake_mcmc.R
\name{drake_mcmc}
\alias{drake_mcmc}
\title{drake_mcmc}
\usage{
drake_mcmc(d, iter = 4000, warmup = iter/2, chains = 3, thin = 1,
  whichmodel = NULL)
}
\arguments{
\item{iter}{number of interations to be run (default=2000)}

\item{chains}{number of chains to be run (default=3)}

\item{thin}{when you want to thin (default=10)}

\item{whichmodel}{characeter of which model you want to run}

\item{dat}{response variable which follows binomial dist}

\item{clusters}{number of clusters to be used (default=nchains)}

\item{burnin}{number of samples to be used as burnin (technically adaption, see link below)}

\item{inits}{expects TRUE/FALSE, if TRUE will use maximum likelihood to get starting values. Needs 3 chains.}
}
\value{
A MCMC object
}
\description{
A wrapper function to make running mcmc fns on drake easier
}
\examples{
drake_mcmc(d, iter = 1000, chains = 3, clusters=3, warmup = iter/2, thin=1, whichmodel=="asg_common")

}
\seealso{
\url{http://www.mikemeredith.net/blog/2016/Adapt_or_burn.htm}
}
