% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_res_df.R
\name{get_res_df}
\alias{get_res_df}
\title{get_res_df}
\usage{
get_res_df(y, x, count, group, mcmc.obj, niter, model.name)
}
\arguments{
\item{y}{response variable used in mcmc}

\item{x}{explanatory variable used in mcmc}

\item{count}{n in binomial dist}

\item{group}{groups of response}

\item{mcmc.obj}{mcmc object}

\item{niter}{number of interations to be ran}

\item{model.name}{name of the model associated with df}
}
\value{
data.frame
}
\description{
Use mcmc objects from package to get posterior estimates of theta
}
\examples{




}
